feat: handle display formatting for scientific notation inputs
Previous logic accounted for cases where the display
was set to to a large value that needed to be converted
to scientific notation. But if the display was set to
a value that was already in scientific notation, it would
not properly format it such that the total number of
digits was equal to TOTAL_DISPLAY_DIGITS. This fix
handles that case.